backend/requirements.txt
azure-functions>=1.15.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
shazamio==0.4.0.1
mutagen==1.47.0
pydantic==1.10.2 